+++ packages/social-metadata-previews/src/shared/ImageSelect.js
